requirements.txt
google-cloud-aiplatform==1.30.1
pip install protobuf==3.20.1
